// src/serverApi/v3/models/school-external-tool-search-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { SchoolExternalToolResponse } from './school-external-tool-response';

/**
 * 
 * @export
 * @interface SchoolExternalToolSearchListResponse
 */
export interface SchoolExternalToolSearchListResponse {
    /**
     * 
     * @type {Array<SchoolExternalToolResponse>}
     * @memberof SchoolExternalToolSearchListResponse
     */
    data: Array<SchoolExternalToolResponse>;
}